MDEV-22970: Disable MDEV-8139 due to corruption concerns
Contrary to our exceptations, it seems that a mini-transaction can
allocate a page that it had freed earlier. The function mtr_t::init()
is not prepared to deal with this, and it could happen that a
newly initialized page will be scrubbed instead. This can affect
the operation on page_compressed tables, or any InnoDB data files when
innodb_background_scrub_data_uncompressed=ON.

Also, buf_read_recv_pages() can interfere with the MDEV-8139 logic
during crash recovery.

Let us temporarily disable MDEV-8139 due to such concerns.

Note: Scrubbing will partially work thanks to MDEV-15528. Only in
cases where the page does not exist in the buffer pool at the time
of the page flush, we would skip the scrubbing action.