Deb: Enable Snappy (for InnoDB and RocksDB)
The build-dependency was already added in 21bbe10bb345.
It should also be defined as a build flag to actually be available.

Also extend Debian autopkgtest to test for all InnoDB compression methods
that are expected to work.

This upstreams the remaining changes of downstream Debian commit
https://salsa.debian.org/mariadb-team/mariadb-10.3/-/commit/278531a7dfa7d60a60b067d089860c92a4e1221b

In addition, simplify Debian autopkgtest RocksDB part and make it more
verbose so that future regressions like this are easier to debug.