MDEV-23304 Insert into TEMPORARY TABLE fails to invoke mtr_t::set_modified()
In MDEV-21724 or possibly already in MDEV-12353 a bug was introduced to
the handling of temporary tables.

Whether or not redo log will be written, mtr_t::set_modified()
must be invoked to register any changes to a page.

page_cur_insert_rec_low(): Invoke mtr_t::set_modified() also
when skipping the redo log write.