MDEV-23438: Assertion `!field->is_null()' failed in my_decimal::my_decimal fails in ONLY_FULL_GROUP_BY mode
The issue here is query with aggregate function and non-aggregate field
in the SELECT LIST with ONLY_FULL_GROUP_BY was not disallowed.
In ONLY_FULL_GROUP_BY mode non-aggregate fields are only allowed inside an
aggregate functions or the non-aggregate fields are part of the GROUP BY clause.

In the query for the failing assert the non-aggregate field was inside
a WINDOW function and the window function was treated as an aggregate function
and so no error was thrown.

The fix would be to make sure to mark that a non-aggregate field is used inside a
window function and not an aggregate function and throw an error then.