MDEV-23399: Use mysql_mutex_t for buf_pool.mutex and flush_list_mutex
We want to avoid the overhead of os_event_t and use condition variables
directly. For this, the two mutexes must be of the regular type.

For now, both mutexes will be declared with the
PTHREAD_MUTEX_ADAPTIVE_NP attribute (MY_MUTEX_INIT_FAST).
The critical sections of buf_pool.flush_list_mutex should be shorter than
those for buf_pool.mutex, because in the worst case, they cover a linear
scan of buf_pool.flush_list, while the worst case of a critical section
of buf_pool.mutex covers a linear scan of the potentially much longer
buf_pool.LRU list.