Work around MDEV-23416 (change Warning to Note)
Ever since MDEV-15053 changed something in the page flushing,
we are occasionally observing pending I/O for a data file that
is about to be deleted.

fil_check_pending_io(): Change the Warning to a note.

This message was already made less frequent in
commit dcc0baf5405b220384b9e1e07d8b9e3ff97b60f4 (10.5.4)
and commit 65f831d17c84900c1faea49164688e2f5ce59563 (10.3.24, 10.4.14).