MDEV-23650 Test S3 in buildbot
Temporarily disable failing S3 tests to allow the suite to run in buildbot:

s3.partition and s3.partition_move are disabled due to MDEV-23648.
While the problem is supposedly generic, the tests don't fail with Amazon
setup, so they are disabled conditionally for emulator (MinIO).

s3.replication_partition is disabled due to MDEV-23730,
it fails both with Amazon and MinIO.

s3.replication_mixed and s3.replication_stmt are disabled due to MDEV-23770,
they also fail both with Amazon and the emulator.