Revert "MDEV-23776 Test encryption.create_or_replace fails with a warning"
This reverts commit e33f7b6faaacf49881fd913115b0735c62aba395.
The change seems to have introduced intermittent failures of the test
encryption.innodb-bad-key-change on many platforms.

The failure that we were trying to address was not reproduced on 10.2.
It could be related to commit a7dd7c899356b2d3a7f79e6ebba5d854ed63ae9d
(MDEV-23651) or de942c9f618b590a01a7960c171d7e50e435708f (MDEV-15983)
or other changes that reduced contention on fil_system.mutex in 10.3.

The fix that we are hereby reverting from 10.2 seems to work fine
on 10.3 and 10.4.