MDEV-23806 Undo page corruption on recovery
When commit 0fd3def284b78dff71590686e8f82571fc3808e1 removed
the MLOG_UNDO_ERASE_END record in MariaDB 10.3.3 in an attempt
to reduce our redo log volume, it introduced technical debt for
commit 56f6dab1d0e5a464ea49c1e5efb0032a0f5cea3e (MDEV-21174) and
commit 7ae21b18a6b73bbc3bf1ff448faf60c29ac1d386 (MDEV-12353),
which optimized mtr_t::write() (née mlog_write_ulint()) so
that the initial bytes that are unchanged are not logged.

trx_undo_report_row_operation(): Write a log record for the
memset() operation if the page is not going to be freed.