Make innodb_buffer_pool_pages_flushed reliable
The function buf_flush_stats() was not called in every code path.
Previously, I thought that it was intentional, but it does not
really make much sense.

Let us increment the statistics at the low level, using a normal
mutex-protected variable. This is similar to
buffer_pool_pages_LRU_flushed, which is also protected by buf_pool.mutex.

Note that show_innodb_vars() directly links to the mutex-protected
variables, and reads of the counters will not be protected by
buf_pool.mutex.