MDEV-21829: Packed Sort Keys in Unique: Address review input
Make the Unique class an interface, and move the implementation
into class Unique_impl.

The idea is to allow seamless change of Unique class's implementation/
to something else.

The construction of Unique object still exposes the implementation:
the code uses "new Unique_impl(...)".  It will be hidden in a factory
function.